preprocess_module/requirements.txt
numpy
torch
transformers
langdetect
streamlit
google-generativeai
